+++ trainers-workload-service/src/main/java/com/trainersworkloadservice/model/YearEntity.java
@@ -41,7 +41,7 @@ public class YearEntity {
 
     @NotNull
     @Column(nullable = false)
-    private Short workYear;
+    private Integer workYear;
 
     @Builder.Default
     @OneToMany(mappedBy = "year", cascade = CascadeType.ALL, orphanRemoval = true, fetch = LAZY)
