+++ trainers-workload-service/src/main/java/com/trainersworkloadservice/repository/MonthRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface MonthRepository extends JpaRepository<MonthEntity,Long> {
-    Optional<MonthEntity> findByYearIdAndMonthOfYear(Long yearId, short month);
+    Optional<MonthEntity> findByYearIdAndMonthOfYear(Long yearId, int month);
 
-    int deleteByYearIdAndMonthOfYear(Long yearId, short month);
+    int deleteByYearIdAndMonthOfYear(Long yearId, int month);
 }
